src/main/java/main/peppa/springframework/recipe/converters/CategoryCommandToCategory.java
package main.peppa.springframework.recipe.converters;

import lombok.Synchronized;
import main.peppa.springframework.recipe.commands.CategoryCommand;
import main.peppa.springframework.recipe.model.Category;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

@Component
public class CategoryCommandToCategory implements Converter<CategoryCommand, Category> {
    @Synchronized
    @Nullable
    @Override
    public Category convert(CategoryCommand source) {
        if (source == null) {
            return null;
        }
        final Category category = new Category();
        category.setId(source.getId());
        category.setDescription(source.getDescription());
        return category;
    }
}
